refactor(tests): reorg and dedup unit tests
