+++ tests/unit/setup.ts
@@ -1,0 +1,13 @@
+jest.mock('@/utils/bee', () => ({
+  ...jest.requireActual('@/utils/bee'),
+  getFeedData: jest.fn(),
+  fetchStamp: jest.fn(),
+}));
+
+jest.mock('@/utils/mantaray', () => ({
+  ...jest.requireActual('@/utils/mantaray'),
+  loadMantaray: jest.fn(),
+  getAllNodeEntries: jest.fn(),
+}));
+
+export {};
